requirements.txt
pygame-ce==2.5.0
zengl==2.4.1